+++ src/server/game/Globals/ObjectMgr.cpp
@@ -3330,7 +3330,7 @@ void ObjectMgr::LoadPlayerCreateLevelStatsData()
 
                 PlayerLevelInfo& levelInfo = info->levelInfo[current_level - 1];
                 for (int i = 0; i < MAX_STATS; i++)
-                    levelInfo.stats[i] = fields[i + 3].GetUInt8();
+                    levelInfo.stats[i] = fields[i + 3].GetUInt16();
             }
 
             ++count;
